# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/balance": {
            "get": {
                "description": "دریافت موجودی کل و قابل برداشت کاربر",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "balance"
                ],
                "summary": "دریافت موجودی حساب",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "شناسه کاربر",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "موجودی حساب",
                        "schema": {
                            "$ref": "#/definitions/models.BalanceResponse"
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/charge": {
            "post": {
                "description": "شارژ حساب کاربری با مبلغ مشخص",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "charge"
                ],
                "summary": "شارژ کردن حساب",
                "parameters": [
                    {
                        "description": "درخواست شارژ",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChargeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "شارژ موفق",
                        "schema": {
                            "$ref": "#/definitions/models.ChargeResponse"
                        }
                    },
                    "400": {
                        "description": "درخواست نامعتبر",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "بررسی سلامت سرویس و اتصال به دیتابیس",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "بررسی وضعیت سرویس",
                "responses": {
                    "200": {
                        "description": "سرویس سالم است",
                        "schema": {
                            "$ref": "#/definitions/models.HealthResponse"
                        }
                    },
                    "500": {
                        "description": "سرویس در وضعیت بد است",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "description": "دریافت لیست تمام تراکنش‌های یک کاربر",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "دریافت تاریخچه تراکنش‌ها",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "شناسه کاربر",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "لیست تراکنش‌ها",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/withdraw": {
            "post": {
                "description": "درخواست برداشت مبلغ از حساب (درخواست معوق می‌شود)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "withdraw"
                ],
                "summary": "درخواست برداشت",
                "parameters": [
                    {
                        "description": "درخواست برداشت",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "درخواست برداشت دریافت شد",
                        "schema": {
                            "$ref": "#/definitions/models.WithdrawResponse"
                        }
                    },
                    "400": {
                        "description": "درخواست نامعتبر",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "خطای سرور",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BalanceResponse": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "withdrawable": {
                    "type": "integer"
                }
            }
        },
        "models.ChargeRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "release_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.ChargeResponse": {
            "type": "object",
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "models.HealthResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "positive for charge, negative for withdraw",
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "release_at": {
                    "description": "optional for charge",
                    "type": "string"
                },
                "status": {
                    "description": "\"pending\", \"completed\", \"failed\"",
                    "type": "string"
                },
                "type": {
                    "description": "\"charge\" or \"withdraw\"",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.WithdrawRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "idempotency_key": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.WithdrawResponse": {
            "type": "object",
            "properties": {
                "idempotency_key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}